pyproject.toml
[tool.poetry]
name = "ai_image_gen"
version = "0.1.4"
description = "AI-powered agent for image processing and generation, as well as video generation from text prompts"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
torch = ">=2.2.0"
diffusers = ">=0.26.0"
transformers = ">=4.37.0"
accelerate = ">=0.27.0"
fastapi = ">=0.109.0"
uvicorn = ">=0.27.0"
pillow = ">=10.2.0"
python-dotenv = ">=1.0.0"
python-multipart = ">=0.0.7"
requests = ">=2.31.0"
runpod = ">=1.5.0"
pydantic = ">=2.5.0"
moviepy = "1.0.3"
boto3 = "^1.37.13"
numpy = "^2.2.4"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
black = ">=23.12.0"
isort = ">=5.13.0"
flake8 = ">=7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]

[tool.ruff]
# Other general settings (e.g., line length, formatting)

[tool.ruff.lint]
ignore = ["E402"]  # Ignore E402 errors